+++ src/main/java/com/example/demo/study/domain/Participant.java
@@ -1,6 +1,13 @@
 package com.example.demo.study.domain;
 
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
 public class Participant {
 
 	private Long id;
+
+	private Long studyId;
 }
